+++ fltk-sys/cfltk/cfl_table.cpp
@@ -380,12 +380,8 @@ void Fl_Table_delete(Fl_Table *self) {
     delete self;
 }
 
-void Fl_Table_set_image_with_size(Fl_Table *self, void *image, int w, int h) {
-    LOCK(self->image(((Fl_Image *)image)->copy(w, h)); self->redraw();)
-}
-
 void Fl_Table_set_image(Fl_Table *self, void *image) {
-    LOCK(self->image(((Fl_Image *)image)->copy()); self->redraw();)
+    LOCK(self->image(((Fl_Image *)image)))
 }
 
 void Fl_Table_set_handler(Fl_Table *self, custom_handler_callback cb, void *data) {
@@ -723,12 +719,8 @@ void Fl_Table_Row_delete(Fl_Table_Row *self) {
     delete self;
 }
 
-void Fl_Table_Row_set_image_with_size(Fl_Table_Row *self, void *image, int w, int h) {
-    LOCK(self->image(((Fl_Image *)image)->copy(w, h)); self->redraw();)
-}
-
 void Fl_Table_Row_set_image(Fl_Table_Row *self, void *image) {
-    LOCK(self->image(((Fl_Image *)image)->copy()); self->redraw();)
+    LOCK(self->image(((Fl_Image *)image)))
 }
 
 void Fl_Table_Row_set_handler(Fl_Table_Row *self, custom_handler_callback cb, void *data) {