+++ fltk-sys/src/browser.rs
@@ -147,14 +147,6 @@ extern "C" {
     pub fn Fl_Widget_set_image(arg1: *mut Fl_Widget, arg2: *mut ::std::os::raw::c_void);
 }
 extern "C" {
-    pub fn Fl_Widget_set_image_with_size(
-        arg1: *mut Fl_Widget,
-        arg2: *mut ::std::os::raw::c_void,
-        arg3: ::std::os::raw::c_int,
-        arg4: ::std::os::raw::c_int,
-    );
-}
-extern "C" {
     pub fn Fl_Widget_set_handler(
         self_: *mut Fl_Widget,
         cb: custom_handler_callback,
@@ -376,14 +368,6 @@ extern "C" {
 }
 extern "C" {
     pub fn Fl_Browser_set_image(arg1: *mut Fl_Browser, arg2: *mut ::std::os::raw::c_void);
-}
-extern "C" {
-    pub fn Fl_Browser_set_image_with_size(
-        arg1: *mut Fl_Browser,
-        arg2: *mut ::std::os::raw::c_void,
-        arg3: ::std::os::raw::c_int,
-        arg4: ::std::os::raw::c_int,
-    );
 }
 extern "C" {
     pub fn Fl_Browser_set_handler(
@@ -706,14 +690,6 @@ extern "C" {
 }
 extern "C" {
     pub fn Fl_Hold_Browser_set_image(arg1: *mut Fl_Hold_Browser, arg2: *mut ::std::os::raw::c_void);
-}
-extern "C" {
-    pub fn Fl_Hold_Browser_set_image_with_size(
-        arg1: *mut Fl_Hold_Browser,
-        arg2: *mut ::std::os::raw::c_void,
-        arg3: ::std::os::raw::c_int,
-        arg4: ::std::os::raw::c_int,
-    );
 }
 extern "C" {
     pub fn Fl_Hold_Browser_set_handler(
@@ -1066,14 +1042,6 @@ extern "C" {
     pub fn Fl_Select_Browser_set_image(
         arg1: *mut Fl_Select_Browser,
         arg2: *mut ::std::os::raw::c_void,
-    );
-}
-extern "C" {
-    pub fn Fl_Select_Browser_set_image_with_size(
-        arg1: *mut Fl_Select_Browser,
-        arg2: *mut ::std::os::raw::c_void,
-        arg3: ::std::os::raw::c_int,
-        arg4: ::std::os::raw::c_int,
     );
 }
 extern "C" {
@@ -1442,14 +1410,6 @@ extern "C" {
     );
 }
 extern "C" {
-    pub fn Fl_Multi_Browser_set_image_with_size(
-        arg1: *mut Fl_Multi_Browser,
-        arg2: *mut ::std::os::raw::c_void,
-        arg3: ::std::os::raw::c_int,
-        arg4: ::std::os::raw::c_int,
-    );
-}
-extern "C" {
     pub fn Fl_Multi_Browser_set_handler(
         self_: *mut Fl_Multi_Browser,
         cb: custom_handler_callback,
@@ -1797,14 +1757,6 @@ extern "C" {
 }
 extern "C" {
     pub fn Fl_File_Browser_set_image(arg1: *mut Fl_File_Browser, arg2: *mut ::std::os::raw::c_void);
-}
-extern "C" {
-    pub fn Fl_File_Browser_set_image_with_size(
-        arg1: *mut Fl_File_Browser,
-        arg2: *mut ::std::os::raw::c_void,
-        arg3: ::std::os::raw::c_int,
-        arg4: ::std::os::raw::c_int,
-    );
 }
 extern "C" {
     pub fn Fl_File_Browser_set_handler(