+++ fltk-sys/src/button.rs
@@ -147,14 +147,6 @@ extern "C" {
     pub fn Fl_Widget_set_image(arg1: *mut Fl_Widget, arg2: *mut ::std::os::raw::c_void);
 }
 extern "C" {
-    pub fn Fl_Widget_set_image_with_size(
-        arg1: *mut Fl_Widget,
-        arg2: *mut ::std::os::raw::c_void,
-        arg3: ::std::os::raw::c_int,
-        arg4: ::std::os::raw::c_int,
-    );
-}
-extern "C" {
     pub fn Fl_Widget_set_handler(
         self_: *mut Fl_Widget,
         cb: custom_handler_callback,
@@ -376,14 +368,6 @@ extern "C" {
 }
 extern "C" {
     pub fn Fl_Button_set_image(arg1: *mut Fl_Button, arg2: *mut ::std::os::raw::c_void);
-}
-extern "C" {
-    pub fn Fl_Button_set_image_with_size(
-        arg1: *mut Fl_Button,
-        arg2: *mut ::std::os::raw::c_void,
-        arg3: ::std::os::raw::c_int,
-        arg4: ::std::os::raw::c_int,
-    );
 }
 extern "C" {
     pub fn Fl_Button_set_handler(
@@ -631,14 +615,6 @@ extern "C" {
 }
 extern "C" {
     pub fn Fl_Check_Button_set_image(arg1: *mut Fl_Check_Button, arg2: *mut ::std::os::raw::c_void);
-}
-extern "C" {
-    pub fn Fl_Check_Button_set_image_with_size(
-        arg1: *mut Fl_Check_Button,
-        arg2: *mut ::std::os::raw::c_void,
-        arg3: ::std::os::raw::c_int,
-        arg4: ::std::os::raw::c_int,
-    );
 }
 extern "C" {
     pub fn Fl_Check_Button_set_handler(
@@ -910,14 +886,6 @@ extern "C" {
 }
 extern "C" {
     pub fn Fl_Radio_Button_set_image(arg1: *mut Fl_Radio_Button, arg2: *mut ::std::os::raw::c_void);
-}
-extern "C" {
-    pub fn Fl_Radio_Button_set_image_with_size(
-        arg1: *mut Fl_Radio_Button,
-        arg2: *mut ::std::os::raw::c_void,
-        arg3: ::std::os::raw::c_int,
-        arg4: ::std::os::raw::c_int,
-    );
 }
 extern "C" {
     pub fn Fl_Radio_Button_set_handler(
@@ -1197,14 +1165,6 @@ extern "C" {
     );
 }
 extern "C" {
-    pub fn Fl_Toggle_Button_set_image_with_size(
-        arg1: *mut Fl_Toggle_Button,
-        arg2: *mut ::std::os::raw::c_void,
-        arg3: ::std::os::raw::c_int,
-        arg4: ::std::os::raw::c_int,
-    );
-}
-extern "C" {
     pub fn Fl_Toggle_Button_set_handler(
         self_: *mut Fl_Toggle_Button,
         cb: custom_handler_callback,
@@ -1480,14 +1440,6 @@ extern "C" {
 }
 extern "C" {
     pub fn Fl_Round_Button_set_image(arg1: *mut Fl_Round_Button, arg2: *mut ::std::os::raw::c_void);
-}
-extern "C" {
-    pub fn Fl_Round_Button_set_image_with_size(
-        arg1: *mut Fl_Round_Button,
-        arg2: *mut ::std::os::raw::c_void,
-        arg3: ::std::os::raw::c_int,
-        arg4: ::std::os::raw::c_int,
-    );
 }
 extern "C" {
     pub fn Fl_Round_Button_set_handler(
@@ -1797,14 +1749,6 @@ extern "C" {
     pub fn Fl_Radio_Round_Button_set_image(
         arg1: *mut Fl_Radio_Round_Button,
         arg2: *mut ::std::os::raw::c_void,
-    );
-}
-extern "C" {
-    pub fn Fl_Radio_Round_Button_set_image_with_size(
-        arg1: *mut Fl_Radio_Round_Button,
-        arg2: *mut ::std::os::raw::c_void,
-        arg3: ::std::os::raw::c_int,
-        arg4: ::std::os::raw::c_int,
     );
 }
 extern "C" {
@@ -2162,14 +2106,6 @@ extern "C" {
     );
 }
 extern "C" {
-    pub fn Fl_Radio_Light_Button_set_image_with_size(
-        arg1: *mut Fl_Radio_Light_Button,
-        arg2: *mut ::std::os::raw::c_void,
-        arg3: ::std::os::raw::c_int,
-        arg4: ::std::os::raw::c_int,
-    );
-}
-extern "C" {
     pub fn Fl_Radio_Light_Button_set_handler(
         self_: *mut Fl_Radio_Light_Button,
         cb: custom_handler_callback,
@@ -2485,14 +2421,6 @@ extern "C" {
     pub fn Fl_Light_Button_set_image(arg1: *mut Fl_Light_Button, arg2: *mut ::std::os::raw::c_void);
 }
 extern "C" {
-    pub fn Fl_Light_Button_set_image_with_size(
-        arg1: *mut Fl_Light_Button,
-        arg2: *mut ::std::os::raw::c_void,
-        arg3: ::std::os::raw::c_int,
-        arg4: ::std::os::raw::c_int,
-    );
-}
-extern "C" {
     pub fn Fl_Light_Button_set_handler(
         self_: *mut Fl_Light_Button,
         cb: custom_handler_callback,
@@ -2770,14 +2698,6 @@ extern "C" {
     );
 }
 extern "C" {
-    pub fn Fl_Repeat_Button_set_image_with_size(
-        arg1: *mut Fl_Repeat_Button,
-        arg2: *mut ::std::os::raw::c_void,
-        arg3: ::std::os::raw::c_int,
-        arg4: ::std::os::raw::c_int,
-    );
-}
-extern "C" {
     pub fn Fl_Repeat_Button_set_handler(
         self_: *mut Fl_Repeat_Button,
         cb: custom_handler_callback,
@@ -3052,14 +2972,6 @@ extern "C" {
     pub fn Fl_Return_Button_set_image(
         arg1: *mut Fl_Return_Button,
         arg2: *mut ::std::os::raw::c_void,
-    );
-}
-extern "C" {
-    pub fn Fl_Return_Button_set_image_with_size(
-        arg1: *mut Fl_Return_Button,
-        arg2: *mut ::std::os::raw::c_void,
-        arg3: ::std::os::raw::c_int,
-        arg4: ::std::os::raw::c_int,
     );
 }
 extern "C" {