+++ fltk-sys/src/group.rs
@@ -147,14 +147,6 @@ extern "C" {
     pub fn Fl_Widget_set_image(arg1: *mut Fl_Widget, arg2: *mut ::std::os::raw::c_void);
 }
 extern "C" {
-    pub fn Fl_Widget_set_image_with_size(
-        arg1: *mut Fl_Widget,
-        arg2: *mut ::std::os::raw::c_void,
-        arg3: ::std::os::raw::c_int,
-        arg4: ::std::os::raw::c_int,
-    );
-}
-extern "C" {
     pub fn Fl_Widget_set_handler(
         self_: *mut Fl_Widget,
         cb: custom_handler_callback,
@@ -376,14 +368,6 @@ extern "C" {
 }
 extern "C" {
     pub fn Fl_Group_set_image(arg1: *mut Fl_Group, arg2: *mut ::std::os::raw::c_void);
-}
-extern "C" {
-    pub fn Fl_Group_set_image_with_size(
-        arg1: *mut Fl_Group,
-        arg2: *mut ::std::os::raw::c_void,
-        arg3: ::std::os::raw::c_int,
-        arg4: ::std::os::raw::c_int,
-    );
 }
 extern "C" {
     pub fn Fl_Group_set_handler(
@@ -646,14 +630,6 @@ extern "C" {
     pub fn Fl_Pack_set_image(arg1: *mut Fl_Pack, arg2: *mut ::std::os::raw::c_void);
 }
 extern "C" {
-    pub fn Fl_Pack_set_image_with_size(
-        arg1: *mut Fl_Pack,
-        arg2: *mut ::std::os::raw::c_void,
-        arg3: ::std::os::raw::c_int,
-        arg4: ::std::os::raw::c_int,
-    );
-}
-extern "C" {
     pub fn Fl_Pack_set_handler(
         self_: *mut Fl_Pack,
         cb: custom_handler_callback,
@@ -912,14 +888,6 @@ extern "C" {
 }
 extern "C" {
     pub fn Fl_Scroll_set_image(arg1: *mut Fl_Scroll, arg2: *mut ::std::os::raw::c_void);
-}
-extern "C" {
-    pub fn Fl_Scroll_set_image_with_size(
-        arg1: *mut Fl_Scroll,
-        arg2: *mut ::std::os::raw::c_void,
-        arg3: ::std::os::raw::c_int,
-        arg4: ::std::os::raw::c_int,
-    );
 }
 extern "C" {
     pub fn Fl_Scroll_set_handler(
@@ -1201,14 +1169,6 @@ extern "C" {
     pub fn Fl_Tabs_set_image(arg1: *mut Fl_Tabs, arg2: *mut ::std::os::raw::c_void);
 }
 extern "C" {
-    pub fn Fl_Tabs_set_image_with_size(
-        arg1: *mut Fl_Tabs,
-        arg2: *mut ::std::os::raw::c_void,
-        arg3: ::std::os::raw::c_int,
-        arg4: ::std::os::raw::c_int,
-    );
-}
-extern "C" {
     pub fn Fl_Tabs_set_handler(
         self_: *mut Fl_Tabs,
         cb: custom_handler_callback,
@@ -1469,14 +1429,6 @@ extern "C" {
     pub fn Fl_Tile_set_image(arg1: *mut Fl_Tile, arg2: *mut ::std::os::raw::c_void);
 }
 extern "C" {
-    pub fn Fl_Tile_set_image_with_size(
-        arg1: *mut Fl_Tile,
-        arg2: *mut ::std::os::raw::c_void,
-        arg3: ::std::os::raw::c_int,
-        arg4: ::std::os::raw::c_int,
-    );
-}
-extern "C" {
     pub fn Fl_Tile_set_handler(
         self_: *mut Fl_Tile,
         cb: custom_handler_callback,
@@ -1735,14 +1687,6 @@ extern "C" {
 }
 extern "C" {
     pub fn Fl_Wizard_set_image(arg1: *mut Fl_Wizard, arg2: *mut ::std::os::raw::c_void);
-}
-extern "C" {
-    pub fn Fl_Wizard_set_image_with_size(
-        arg1: *mut Fl_Wizard,
-        arg2: *mut ::std::os::raw::c_void,
-        arg3: ::std::os::raw::c_int,
-        arg4: ::std::os::raw::c_int,
-    );
 }
 extern "C" {
     pub fn Fl_Wizard_set_handler(
@@ -2029,14 +1973,6 @@ extern "C" {
     pub fn Fl_Color_Chooser_set_image(
         arg1: *mut Fl_Color_Chooser,
         arg2: *mut ::std::os::raw::c_void,
-    );
-}
-extern "C" {
-    pub fn Fl_Color_Chooser_set_image_with_size(
-        arg1: *mut Fl_Color_Chooser,
-        arg2: *mut ::std::os::raw::c_void,
-        arg3: ::std::os::raw::c_int,
-        arg4: ::std::os::raw::c_int,
     );
 }
 extern "C" {