+++ fltk-sys/src/input.rs
@@ -147,14 +147,6 @@ extern "C" {
     pub fn Fl_Widget_set_image(arg1: *mut Fl_Widget, arg2: *mut ::std::os::raw::c_void);
 }
 extern "C" {
-    pub fn Fl_Widget_set_image_with_size(
-        arg1: *mut Fl_Widget,
-        arg2: *mut ::std::os::raw::c_void,
-        arg3: ::std::os::raw::c_int,
-        arg4: ::std::os::raw::c_int,
-    );
-}
-extern "C" {
     pub fn Fl_Widget_set_handler(
         self_: *mut Fl_Widget,
         cb: custom_handler_callback,
@@ -376,14 +368,6 @@ extern "C" {
 }
 extern "C" {
     pub fn Fl_Input_set_image(arg1: *mut Fl_Input, arg2: *mut ::std::os::raw::c_void);
-}
-extern "C" {
-    pub fn Fl_Input_set_image_with_size(
-        arg1: *mut Fl_Input,
-        arg2: *mut ::std::os::raw::c_void,
-        arg3: ::std::os::raw::c_int,
-        arg4: ::std::os::raw::c_int,
-    );
 }
 extern "C" {
     pub fn Fl_Input_set_handler(
@@ -706,14 +690,6 @@ extern "C" {
 }
 extern "C" {
     pub fn Fl_Int_Input_set_image(arg1: *mut Fl_Int_Input, arg2: *mut ::std::os::raw::c_void);
-}
-extern "C" {
-    pub fn Fl_Int_Input_set_image_with_size(
-        arg1: *mut Fl_Int_Input,
-        arg2: *mut ::std::os::raw::c_void,
-        arg3: ::std::os::raw::c_int,
-        arg4: ::std::os::raw::c_int,
-    );
 }
 extern "C" {
     pub fn Fl_Int_Input_set_handler(
@@ -1045,14 +1021,6 @@ extern "C" {
 }
 extern "C" {
     pub fn Fl_Float_Input_set_image(arg1: *mut Fl_Float_Input, arg2: *mut ::std::os::raw::c_void);
-}
-extern "C" {
-    pub fn Fl_Float_Input_set_image_with_size(
-        arg1: *mut Fl_Float_Input,
-        arg2: *mut ::std::os::raw::c_void,
-        arg3: ::std::os::raw::c_int,
-        arg4: ::std::os::raw::c_int,
-    );
 }
 extern "C" {
     pub fn Fl_Float_Input_set_handler(
@@ -1413,14 +1381,6 @@ extern "C" {
     pub fn Fl_Multiline_Input_set_image(
         arg1: *mut Fl_Multiline_Input,
         arg2: *mut ::std::os::raw::c_void,
-    );
-}
-extern "C" {
-    pub fn Fl_Multiline_Input_set_image_with_size(
-        arg1: *mut Fl_Multiline_Input,
-        arg2: *mut ::std::os::raw::c_void,
-        arg3: ::std::os::raw::c_int,
-        arg4: ::std::os::raw::c_int,
     );
 }
 extern "C" {
@@ -1818,14 +1778,6 @@ extern "C" {
     pub fn Fl_Secret_Input_set_image(arg1: *mut Fl_Secret_Input, arg2: *mut ::std::os::raw::c_void);
 }
 extern "C" {
-    pub fn Fl_Secret_Input_set_image_with_size(
-        arg1: *mut Fl_Secret_Input,
-        arg2: *mut ::std::os::raw::c_void,
-        arg3: ::std::os::raw::c_int,
-        arg4: ::std::os::raw::c_int,
-    );
-}
-extern "C" {
     pub fn Fl_Secret_Input_set_handler(
         self_: *mut Fl_Secret_Input,
         cb: custom_handler_callback,
@@ -2161,14 +2113,6 @@ extern "C" {
 }
 extern "C" {
     pub fn Fl_File_Input_set_image(arg1: *mut Fl_File_Input, arg2: *mut ::std::os::raw::c_void);
-}
-extern "C" {
-    pub fn Fl_File_Input_set_image_with_size(
-        arg1: *mut Fl_File_Input,
-        arg2: *mut ::std::os::raw::c_void,
-        arg3: ::std::os::raw::c_int,
-        arg4: ::std::os::raw::c_int,
-    );
 }
 extern "C" {
     pub fn Fl_File_Input_set_handler(