+++ fltk-sys/src/misc.rs
@@ -147,14 +147,6 @@ extern "C" {
     pub fn Fl_Widget_set_image(arg1: *mut Fl_Widget, arg2: *mut ::std::os::raw::c_void);
 }
 extern "C" {
-    pub fn Fl_Widget_set_image_with_size(
-        arg1: *mut Fl_Widget,
-        arg2: *mut ::std::os::raw::c_void,
-        arg3: ::std::os::raw::c_int,
-        arg4: ::std::os::raw::c_int,
-    );
-}
-extern "C" {
     pub fn Fl_Widget_set_handler(
         self_: *mut Fl_Widget,
         cb: custom_handler_callback,
@@ -376,14 +368,6 @@ extern "C" {
 }
 extern "C" {
     pub fn Fl_Spinner_set_image(arg1: *mut Fl_Spinner, arg2: *mut ::std::os::raw::c_void);
-}
-extern "C" {
-    pub fn Fl_Spinner_set_image_with_size(
-        arg1: *mut Fl_Spinner,
-        arg2: *mut ::std::os::raw::c_void,
-        arg3: ::std::os::raw::c_int,
-        arg4: ::std::os::raw::c_int,
-    );
 }
 extern "C" {
     pub fn Fl_Spinner_set_handler(
@@ -654,14 +638,6 @@ extern "C" {
     pub fn Fl_Clock_set_image(arg1: *mut Fl_Clock, arg2: *mut ::std::os::raw::c_void);
 }
 extern "C" {
-    pub fn Fl_Clock_set_image_with_size(
-        arg1: *mut Fl_Clock,
-        arg2: *mut ::std::os::raw::c_void,
-        arg3: ::std::os::raw::c_int,
-        arg4: ::std::os::raw::c_int,
-    );
-}
-extern "C" {
     pub fn Fl_Clock_set_handler(
         self_: *mut Fl_Clock,
         cb: custom_handler_callback,
@@ -883,14 +859,6 @@ extern "C" {
 }
 extern "C" {
     pub fn Fl_Chart_set_image(arg1: *mut Fl_Chart, arg2: *mut ::std::os::raw::c_void);
-}
-extern "C" {
-    pub fn Fl_Chart_set_image_with_size(
-        arg1: *mut Fl_Chart,
-        arg2: *mut ::std::os::raw::c_void,
-        arg3: ::std::os::raw::c_int,
-        arg4: ::std::os::raw::c_int,
-    );
 }
 extern "C" {
     pub fn Fl_Chart_set_handler(
@@ -1186,14 +1154,6 @@ extern "C" {
 }
 extern "C" {
     pub fn Fl_Progress_set_image(arg1: *mut Fl_Progress, arg2: *mut ::std::os::raw::c_void);
-}
-extern "C" {
-    pub fn Fl_Progress_set_image_with_size(
-        arg1: *mut Fl_Progress,
-        arg2: *mut ::std::os::raw::c_void,
-        arg3: ::std::os::raw::c_int,
-        arg4: ::std::os::raw::c_int,
-    );
 }
 extern "C" {
     pub fn Fl_Progress_set_handler(