+++ fltk-sys/src/text.rs
@@ -147,14 +147,6 @@ extern "C" {
     pub fn Fl_Widget_set_image(arg1: *mut Fl_Widget, arg2: *mut ::std::os::raw::c_void);
 }
 extern "C" {
-    pub fn Fl_Widget_set_image_with_size(
-        arg1: *mut Fl_Widget,
-        arg2: *mut ::std::os::raw::c_void,
-        arg3: ::std::os::raw::c_int,
-        arg4: ::std::os::raw::c_int,
-    );
-}
-extern "C" {
     pub fn Fl_Widget_set_handler(
         self_: *mut Fl_Widget,
         cb: custom_handler_callback,
@@ -586,14 +578,6 @@ extern "C" {
 }
 extern "C" {
     pub fn Fl_Text_Display_set_image(arg1: *mut Fl_Text_Display, arg2: *mut ::std::os::raw::c_void);
-}
-extern "C" {
-    pub fn Fl_Text_Display_set_image_with_size(
-        arg1: *mut Fl_Text_Display,
-        arg2: *mut ::std::os::raw::c_void,
-        arg3: ::std::os::raw::c_int,
-        arg4: ::std::os::raw::c_int,
-    );
 }
 extern "C" {
     pub fn Fl_Text_Display_set_handler(
@@ -1089,14 +1073,6 @@ extern "C" {
 }
 extern "C" {
     pub fn Fl_Text_Editor_set_image(arg1: *mut Fl_Text_Editor, arg2: *mut ::std::os::raw::c_void);
-}
-extern "C" {
-    pub fn Fl_Text_Editor_set_image_with_size(
-        arg1: *mut Fl_Text_Editor,
-        arg2: *mut ::std::os::raw::c_void,
-        arg3: ::std::os::raw::c_int,
-        arg4: ::std::os::raw::c_int,
-    );
 }
 extern "C" {
     pub fn Fl_Text_Editor_set_handler(
@@ -1616,14 +1592,6 @@ extern "C" {
     pub fn Fl_Simple_Terminal_set_image(
         arg1: *mut Fl_Simple_Terminal,
         arg2: *mut ::std::os::raw::c_void,
-    );
-}
-extern "C" {
-    pub fn Fl_Simple_Terminal_set_image_with_size(
-        arg1: *mut Fl_Simple_Terminal,
-        arg2: *mut ::std::os::raw::c_void,
-        arg3: ::std::os::raw::c_int,
-        arg4: ::std::os::raw::c_int,
     );
 }
 extern "C" {