+++ fltk-sys/src/tree.rs
@@ -147,14 +147,6 @@ extern "C" {
     pub fn Fl_Widget_set_image(arg1: *mut Fl_Widget, arg2: *mut ::std::os::raw::c_void);
 }
 extern "C" {
-    pub fn Fl_Widget_set_image_with_size(
-        arg1: *mut Fl_Widget,
-        arg2: *mut ::std::os::raw::c_void,
-        arg3: ::std::os::raw::c_int,
-        arg4: ::std::os::raw::c_int,
-    );
-}
-extern "C" {
     pub fn Fl_Widget_set_handler(
         self_: *mut Fl_Widget,
         cb: custom_handler_callback,
@@ -376,14 +368,6 @@ extern "C" {
 }
 extern "C" {
     pub fn Fl_Tree_set_image(arg1: *mut Fl_Tree, arg2: *mut ::std::os::raw::c_void);
-}
-extern "C" {
-    pub fn Fl_Tree_set_image_with_size(
-        arg1: *mut Fl_Tree,
-        arg2: *mut ::std::os::raw::c_void,
-        arg3: ::std::os::raw::c_int,
-        arg4: ::std::os::raw::c_int,
-    );
 }
 extern "C" {
     pub fn Fl_Tree_set_handler(