+++ fltk-sys/src/valuator.rs
@@ -147,14 +147,6 @@ extern "C" {
     pub fn Fl_Widget_set_image(arg1: *mut Fl_Widget, arg2: *mut ::std::os::raw::c_void);
 }
 extern "C" {
-    pub fn Fl_Widget_set_image_with_size(
-        arg1: *mut Fl_Widget,
-        arg2: *mut ::std::os::raw::c_void,
-        arg3: ::std::os::raw::c_int,
-        arg4: ::std::os::raw::c_int,
-    );
-}
-extern "C" {
     pub fn Fl_Widget_set_handler(
         self_: *mut Fl_Widget,
         cb: custom_handler_callback,
@@ -376,14 +368,6 @@ extern "C" {
 }
 extern "C" {
     pub fn Fl_Slider_set_image(arg1: *mut Fl_Slider, arg2: *mut ::std::os::raw::c_void);
-}
-extern "C" {
-    pub fn Fl_Slider_set_image_with_size(
-        arg1: *mut Fl_Slider,
-        arg2: *mut ::std::os::raw::c_void,
-        arg3: ::std::os::raw::c_int,
-        arg4: ::std::os::raw::c_int,
-    );
 }
 extern "C" {
     pub fn Fl_Slider_set_handler(
@@ -662,14 +646,6 @@ extern "C" {
 }
 extern "C" {
     pub fn Fl_Nice_Slider_set_image(arg1: *mut Fl_Nice_Slider, arg2: *mut ::std::os::raw::c_void);
-}
-extern "C" {
-    pub fn Fl_Nice_Slider_set_image_with_size(
-        arg1: *mut Fl_Nice_Slider,
-        arg2: *mut ::std::os::raw::c_void,
-        arg3: ::std::os::raw::c_int,
-        arg4: ::std::os::raw::c_int,
-    );
 }
 extern "C" {
     pub fn Fl_Nice_Slider_set_handler(
@@ -959,14 +935,6 @@ extern "C" {
     pub fn Fl_Counter_set_image(arg1: *mut Fl_Counter, arg2: *mut ::std::os::raw::c_void);
 }
 extern "C" {
-    pub fn Fl_Counter_set_image_with_size(
-        arg1: *mut Fl_Counter,
-        arg2: *mut ::std::os::raw::c_void,
-        arg3: ::std::os::raw::c_int,
-        arg4: ::std::os::raw::c_int,
-    );
-}
-extern "C" {
     pub fn Fl_Counter_set_handler(
         self_: *mut Fl_Counter,
         cb: custom_handler_callback,
@@ -1242,14 +1210,6 @@ extern "C" {
     pub fn Fl_Dial_set_image(arg1: *mut Fl_Dial, arg2: *mut ::std::os::raw::c_void);
 }
 extern "C" {
-    pub fn Fl_Dial_set_image_with_size(
-        arg1: *mut Fl_Dial,
-        arg2: *mut ::std::os::raw::c_void,
-        arg3: ::std::os::raw::c_int,
-        arg4: ::std::os::raw::c_int,
-    );
-}
-extern "C" {
     pub fn Fl_Dial_set_handler(
         self_: *mut Fl_Dial,
         cb: custom_handler_callback,
@@ -1519,14 +1479,6 @@ extern "C" {
 }
 extern "C" {
     pub fn Fl_Line_Dial_set_image(arg1: *mut Fl_Line_Dial, arg2: *mut ::std::os::raw::c_void);
-}
-extern "C" {
-    pub fn Fl_Line_Dial_set_image_with_size(
-        arg1: *mut Fl_Line_Dial,
-        arg2: *mut ::std::os::raw::c_void,
-        arg3: ::std::os::raw::c_int,
-        arg4: ::std::os::raw::c_int,
-    );
 }
 extern "C" {
     pub fn Fl_Line_Dial_set_handler(
@@ -1807,14 +1759,6 @@ extern "C" {
     pub fn Fl_Roller_set_image(arg1: *mut Fl_Roller, arg2: *mut ::std::os::raw::c_void);
 }
 extern "C" {
-    pub fn Fl_Roller_set_image_with_size(
-        arg1: *mut Fl_Roller,
-        arg2: *mut ::std::os::raw::c_void,
-        arg3: ::std::os::raw::c_int,
-        arg4: ::std::os::raw::c_int,
-    );
-}
-extern "C" {
     pub fn Fl_Roller_set_handler(
         self_: *mut Fl_Roller,
         cb: custom_handler_callback,
@@ -2085,14 +2029,6 @@ extern "C" {
 }
 extern "C" {
     pub fn Fl_Scrollbar_set_image(arg1: *mut Fl_Scrollbar, arg2: *mut ::std::os::raw::c_void);
-}
-extern "C" {
-    pub fn Fl_Scrollbar_set_image_with_size(
-        arg1: *mut Fl_Scrollbar,
-        arg2: *mut ::std::os::raw::c_void,
-        arg3: ::std::os::raw::c_int,
-        arg4: ::std::os::raw::c_int,
-    );
 }
 extern "C" {
     pub fn Fl_Scrollbar_set_handler(
@@ -2380,14 +2316,6 @@ extern "C" {
 }
 extern "C" {
     pub fn Fl_Value_Slider_set_image(arg1: *mut Fl_Value_Slider, arg2: *mut ::std::os::raw::c_void);
-}
-extern "C" {
-    pub fn Fl_Value_Slider_set_image_with_size(
-        arg1: *mut Fl_Value_Slider,
-        arg2: *mut ::std::os::raw::c_void,
-        arg3: ::std::os::raw::c_int,
-        arg4: ::std::os::raw::c_int,
-    );
 }
 extern "C" {
     pub fn Fl_Value_Slider_set_handler(
@@ -2683,14 +2611,6 @@ extern "C" {
     pub fn Fl_Adjuster_set_image(arg1: *mut Fl_Adjuster, arg2: *mut ::std::os::raw::c_void);
 }
 extern "C" {
-    pub fn Fl_Adjuster_set_image_with_size(
-        arg1: *mut Fl_Adjuster,
-        arg2: *mut ::std::os::raw::c_void,
-        arg3: ::std::os::raw::c_int,
-        arg4: ::std::os::raw::c_int,
-    );
-}
-extern "C" {
     pub fn Fl_Adjuster_set_handler(
         self_: *mut Fl_Adjuster,
         cb: custom_handler_callback,
@@ -2970,14 +2890,6 @@ extern "C" {
 }
 extern "C" {
     pub fn Fl_Value_Input_set_image(arg1: *mut Fl_Value_Input, arg2: *mut ::std::os::raw::c_void);
-}
-extern "C" {
-    pub fn Fl_Value_Input_set_image_with_size(
-        arg1: *mut Fl_Value_Input,
-        arg2: *mut ::std::os::raw::c_void,
-        arg3: ::std::os::raw::c_int,
-        arg4: ::std::os::raw::c_int,
-    );
 }
 extern "C" {
     pub fn Fl_Value_Input_set_handler(
@@ -3274,14 +3186,6 @@ extern "C" {
 }
 extern "C" {
     pub fn Fl_Value_Output_set_image(arg1: *mut Fl_Value_Output, arg2: *mut ::std::os::raw::c_void);
-}
-extern "C" {
-    pub fn Fl_Value_Output_set_image_with_size(
-        arg1: *mut Fl_Value_Output,
-        arg2: *mut ::std::os::raw::c_void,
-        arg3: ::std::os::raw::c_int,
-        arg4: ::std::os::raw::c_int,
-    );
 }
 extern "C" {
     pub fn Fl_Value_Output_set_handler(
@@ -3583,14 +3487,6 @@ extern "C" {
     pub fn Fl_Fill_Slider_set_image(arg1: *mut Fl_Fill_Slider, arg2: *mut ::std::os::raw::c_void);
 }
 extern "C" {
-    pub fn Fl_Fill_Slider_set_image_with_size(
-        arg1: *mut Fl_Fill_Slider,
-        arg2: *mut ::std::os::raw::c_void,
-        arg3: ::std::os::raw::c_int,
-        arg4: ::std::os::raw::c_int,
-    );
-}
-extern "C" {
     pub fn Fl_Fill_Slider_set_handler(
         self_: *mut Fl_Fill_Slider,
         cb: custom_handler_callback,
@@ -3878,14 +3774,6 @@ extern "C" {
     pub fn Fl_Fill_Dial_set_image(arg1: *mut Fl_Fill_Dial, arg2: *mut ::std::os::raw::c_void);
 }
 extern "C" {
-    pub fn Fl_Fill_Dial_set_image_with_size(
-        arg1: *mut Fl_Fill_Dial,
-        arg2: *mut ::std::os::raw::c_void,
-        arg3: ::std::os::raw::c_int,
-        arg4: ::std::os::raw::c_int,
-    );
-}
-extern "C" {
     pub fn Fl_Fill_Dial_set_handler(
         self_: *mut Fl_Fill_Dial,
         cb: custom_handler_callback,
@@ -4162,14 +4050,6 @@ extern "C" {
 }
 extern "C" {
     pub fn Fl_Hor_Slider_set_image(arg1: *mut Fl_Hor_Slider, arg2: *mut ::std::os::raw::c_void);
-}
-extern "C" {
-    pub fn Fl_Hor_Slider_set_image_with_size(
-        arg1: *mut Fl_Hor_Slider,
-        arg2: *mut ::std::os::raw::c_void,
-        arg3: ::std::os::raw::c_int,
-        arg4: ::std::os::raw::c_int,
-    );
 }
 extern "C" {
     pub fn Fl_Hor_Slider_set_handler(
@@ -4480,14 +4360,6 @@ extern "C" {
     pub fn Fl_Hor_Fill_Slider_set_image(
         arg1: *mut Fl_Hor_Fill_Slider,
         arg2: *mut ::std::os::raw::c_void,
-    );
-}
-extern "C" {
-    pub fn Fl_Hor_Fill_Slider_set_image_with_size(
-        arg1: *mut Fl_Hor_Fill_Slider,
-        arg2: *mut ::std::os::raw::c_void,
-        arg3: ::std::os::raw::c_int,
-        arg4: ::std::os::raw::c_int,
     );
 }
 extern "C" {
@@ -4844,14 +4716,6 @@ extern "C" {
     pub fn Fl_Hor_Nice_Slider_set_image(
         arg1: *mut Fl_Hor_Nice_Slider,
         arg2: *mut ::std::os::raw::c_void,
-    );
-}
-extern "C" {
-    pub fn Fl_Hor_Nice_Slider_set_image_with_size(
-        arg1: *mut Fl_Hor_Nice_Slider,
-        arg2: *mut ::std::os::raw::c_void,
-        arg3: ::std::os::raw::c_int,
-        arg4: ::std::os::raw::c_int,
     );
 }
 extern "C" {
@@ -5214,14 +5078,6 @@ extern "C" {
     pub fn Fl_Hor_Value_Slider_set_image(
         arg1: *mut Fl_Hor_Value_Slider,
         arg2: *mut ::std::os::raw::c_void,
-    );
-}
-extern "C" {
-    pub fn Fl_Hor_Value_Slider_set_image_with_size(
-        arg1: *mut Fl_Hor_Value_Slider,
-        arg2: *mut ::std::os::raw::c_void,
-        arg3: ::std::os::raw::c_int,
-        arg4: ::std::os::raw::c_int,
     );
 }
 extern "C" {